LLAMA2 and other API refinements
* made it possible to pass temperature to ask-for-list
* refined print method for JSON
* implemented a macro expansion to define all chat and embed functions for
llama2 and OpenAI.  This made it simpler to ensure that there was a common signature
for llama2:embed and gpt:embed, llama2:ask-chat and gpt:ask-chat etc.
The unification of signatures is a step toward making the upstream magic predicates in Allegrograph
compatible with both LLAMA2 and OpenAI models (and presumably new LLMs addded in the future).
* Added default values for all remaining chat/ask-chat/embed parameters in OpenAI and LLAMA2 APIs.
* Added in changes to ask-my-documents and new ask-for-map.
* Reworked macro key-args-fun.
* Also comment out code useful for debugging only.
* acl-llm module in agraph source not acl

Change-Id: I7bd0cf946ed5254c2c8cc03ec51a277e065c6e1c
Reviewed-on: https://gerrit.franz.com:9080/c/acl-llm/+/23904
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Richard Wallace <rsw@franz.com>
Ptest: Hudson <hudson@franz.com>
Tested-by: Hudson <hudson@franz.com>