AG-1726 - support the saving embeding length in vendor
Embedding code needs to know the length of an embedding vector
so we save that information in a vendor object supporting
embedding.

This change is an intermediate step in the evolution of
the clos based llm system.  It's necessary to get dependent
changes to work and will be updated to the final state
in another change.

%ReleaseNote:   no
%Documentation: no

Change-Id: Id0fd39131dc64bd5a7c77950f9763c34a045e37c
Reviewed-on: https://gerrit.franz.com:9080/c/acl-llm/+/27345
Reviewed-by: John Foderaro <jkf@franz.com>
Ptest: Hudson <hudson@franz.com>
Tested-by: Hudson <hudson@franz.com>